chore: add hook
